+++ common/src/main/java/net/caffeinemc/mods/lithium/mixin/util/initialization/package-info.java
@@ -1,0 +1,4 @@
+@MixinConfigOption(description = "Platform dependent initialization hooks.")
+package net.caffeinemc.mods.lithium.mixin.util.initialization;
+
+import net.caffeinemc.gradle.MixinConfigOption;